refactor: move scatter chart to its own component
